feat(radar): move stories to the dedicated storybook workspace
