+++ storybook/package.json
@@ -29,11 +29,13 @@
     "@nivo/marimekko": "workspace:*",
     "@nivo/network": "workspace:*",
     "@nivo/pie": "workspace:*",
+    "@nivo/radar": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@nivo/waffle": "workspace:*",
     "@babel/preset-env": "^7.21.5",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.21.5",
+    "@react-spring/web": "9.4.5 || ^9.7.2",
     "@storybook/addon-essentials": "^7.0.7",
     "@storybook/addon-interactions": "^7.0.7",
     "@storybook/addon-links": "^7.0.7",
